initial CreateTableGenerator and TableRenameGenerator
