+++ sqlitelib/src/main/java/com/fsryan/forsuredb/sqlitelib/diff/TableRenameGenerator.java
@@ -1,0 +1,21 @@
+package com.fsryan.forsuredb.sqlitelib.diff;
+
+import javax.annotation.Nonnull;
+import java.util.Collections;
+import java.util.List;
+
+public class TableRenameGenerator {
+
+    private final String prevName;
+    private final String currName;
+
+    TableRenameGenerator(@Nonnull String prevName, @Nonnull String currName) {
+        this.prevName = prevName;
+        this.currName = currName;
+    }
+
+    @Nonnull
+    public List<String> statements() {
+        return Collections.singletonList("ALTER TABLE " + prevName + " RENAME TO " + currName);
+    }
+}
